requirements.txt
flask>=3.0,<4
flask-cors>=4.0,<5
openai>=1.30,<2
python-dotenv>=1.0,<2
azure-cognitiveservices-speech>=1.38,<2
